Add new module for extracting the density and gsd
The ground sample distance (GSD) is the real distance between two
consecutive pixels. Firstly, it is obtained the altitude of the drone
from the image properties. Secondly, the FOV is calculated using the f
and sensor size parameters. Lastly, basic geometry equations are used
for getting the GSD value.

With the GSD it can be obtained the real image area and, altogether
with the number of pumpkins, the field density can be returned.